refactor: extract chat/completions/anthropic routes from server.py (Phase 2)
server.py 3106 → 1127 lines (-63.6%). Extract 3 core routes and shared
helpers into dedicated modules:

- service/helpers.py: _resolve_*, get_engine, _disconnect_guard, etc.
- routes/completions.py: /v1/completions
- routes/chat.py: /v1/chat/completions + stream_chat_completion
- routes/anthropic.py: /v1/messages + _stream_anthropic_messages

All symbols re-exported from server.py for backward compatibility.
2159 tests pass, 0 regressions. Closes #140 Phase 2.
