+++ vllm_mlx/routes/completions.py
@@ -1,0 +1,153 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Text completion endpoints — /v1/completions."""
+
+import json
+import logging
+import time
+import uuid
+from collections.abc import AsyncIterator
+
+from fastapi import APIRouter, Depends, Request
+from fastapi.responses import Response, StreamingResponse
+
+from ..api.models import (
+    CompletionChoice,
+    CompletionRequest,
+    CompletionResponse,
+    Usage,
+)
+from ..config import get_config
+from ..middleware.auth import check_rate_limit, verify_api_key
+from ..service.helpers import (
+    _disconnect_guard,
+    _resolve_max_tokens,
+    _resolve_model_name,
+    _resolve_temperature,
+    _resolve_top_p,
+    _validate_model_name,
+    _wait_with_disconnect,
+    get_engine,
+    get_usage,
+)
+
+logger = logging.getLogger(__name__)
+
+router = APIRouter()
+
+
+@router.post(
+    "/v1/completions",
+    dependencies=[Depends(verify_api_key), Depends(check_rate_limit)],
+)
+async def create_completion(request: CompletionRequest, raw_request: Request):
+    """Create a text completion."""
+    _validate_model_name(request.model)
+    engine = get_engine(request.model)
+
+    # Handle single prompt or list of prompts
+    prompts = request.prompt if isinstance(request.prompt, list) else [request.prompt]
+
+    # --- Detailed request logging ---
+    prompt_preview = prompts[0][:200] if prompts else "(empty)"
+    prompt_len = sum(len(p) for p in prompts)
+    logger.info(
+        f"[REQUEST] POST /v1/completions stream={request.stream} "
+        f"max_tokens={request.max_tokens} temp={request.temperature} "
+        f"prompt_chars={prompt_len} prompt_preview={prompt_preview!r}"
+    )
+
+    if request.stream:
+        return StreamingResponse(
+            _disconnect_guard(
+                stream_completion(engine, prompts[0], request),
+                raw_request,
+            ),
+            media_type="text/event-stream",
+        )
+
+    # Non-streaming response with timing and timeout
+    start_time = time.perf_counter()
+    timeout = request.timeout or get_config().default_timeout
+    choices = []
+    total_completion_tokens = 0
+    total_prompt_tokens = 0
+
+    for i, prompt in enumerate(prompts):
+        output = await _wait_with_disconnect(
+            engine.generate(
+                prompt=prompt,
+                max_tokens=_resolve_max_tokens(request.max_tokens),
+                temperature=_resolve_temperature(request.temperature),
+                top_p=_resolve_top_p(request.top_p),
+                stop=request.stop,
+            ),
+            raw_request,
+            timeout=timeout,
+        )
+        if output is None:
+            return Response(status_code=499)  # Client closed request
+
+        choices.append(
+            CompletionChoice(
+                index=i,
+                text=output.text,
+                finish_reason=output.finish_reason,
+            )
+        )
+        total_completion_tokens += output.completion_tokens
+        total_prompt_tokens += (
+            output.prompt_tokens if hasattr(output, "prompt_tokens") else 0
+        )
+
+    elapsed = time.perf_counter() - start_time
+    tokens_per_sec = total_completion_tokens / elapsed if elapsed > 0 else 0
+    logger.info(
+        f"Completion: {total_prompt_tokens} prompt + {total_completion_tokens} completion tokens in {elapsed:.2f}s ({tokens_per_sec:.1f} tok/s)"
+    )
+
+    comp_response = CompletionResponse(
+        model=_resolve_model_name(request.model),
+        choices=choices,
+        usage=Usage(
+            prompt_tokens=total_prompt_tokens,
+            completion_tokens=total_completion_tokens,
+            total_tokens=total_prompt_tokens + total_completion_tokens,
+        ),
+    )
+    return Response(
+        content=comp_response.model_dump_json(exclude_none=True),
+        media_type="application/json",
+    )
+
+
+async def stream_completion(
+    engine,
+    prompt: str,
+    request: CompletionRequest,
+) -> AsyncIterator[str]:
+    """Stream completion response."""
+    async for output in engine.stream_generate(
+        prompt=prompt,
+        max_tokens=_resolve_max_tokens(request.max_tokens),
+        temperature=_resolve_temperature(request.temperature),
+        top_p=_resolve_top_p(request.top_p),
+        stop=request.stop,
+    ):
+        data = {
+            "id": f"cmpl-{uuid.uuid4().hex[:8]}",
+            "object": "text_completion",
+            "created": int(time.time()),
+            "model": _resolve_model_name(request.model),
+            "choices": [
+                {
+                    "index": 0,
+                    "text": output.new_text,
+                    "finish_reason": output.finish_reason if output.finished else None,
+                }
+            ],
+        }
+        if output.finished:
+            data["usage"] = get_usage(output).model_dump()
+        yield f"data: {json.dumps(data)}\n\n"
+
+    yield "data: [DONE]\n\n"
