+++ vllm_mlx/server.py
@@ -40,77 +40,92 @@ The server provides:
 import argparse
 import asyncio
 import gc
-import hashlib
-import json
 import logging
 import os
-import time
-import uuid
-from collections.abc import AsyncIterator
 
 import uvicorn
-from fastapi import Depends, FastAPI, HTTPException, Request
+from fastapi import FastAPI, Request
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import Response, StreamingResponse
 
-# Import from new modular API
 # Re-export for backwards compatibility with tests
-from .api.anthropic_adapter import anthropic_to_openai, openai_to_anthropic
-from .api.anthropic_models import AnthropicRequest
+from .api.anthropic_adapter import (  # noqa: F401
+    anthropic_to_openai,
+    openai_to_anthropic,
+)
+from .api.anthropic_models import AnthropicRequest  # noqa: F401
 from .api.models import (
     AssistantMessage,  # noqa: F401
     ChatCompletionChoice,  # noqa: F401
     ChatCompletionChunk,  # noqa: F401
     ChatCompletionChunkChoice,  # noqa: F401
     ChatCompletionChunkDelta,  # noqa: F401
-    ChatCompletionRequest,
-    ChatCompletionResponse,
-    ChoiceLogProbs,
+    ChatCompletionRequest,  # noqa: F401
+    ChatCompletionResponse,  # noqa: F401
+    ChoiceLogProbs,  # noqa: F401
     CompletionChoice,  # noqa: F401
-    CompletionRequest,
-    CompletionResponse,
+    CompletionRequest,  # noqa: F401
+    CompletionResponse,  # noqa: F401
     CompletionTokensDetails,  # noqa: F401
     ContentPart,  # noqa: F401
-    FunctionCall,
+    FunctionCall,  # noqa: F401
     ImageUrl,  # noqa: F401
     MCPServerInfo,  # noqa: F401
     MCPToolInfo,  # noqa: F401
     Message,  # noqa: F401
     ModelInfo,  # noqa: F401
-    TokenLogProb,
-    ToolCall,
-    TopLogProb,
+    TokenLogProb,  # noqa: F401
+    ToolCall,  # noqa: F401
+    TopLogProb,  # noqa: F401
     Usage,  # noqa: F401
     VideoUrl,  # noqa: F401
 )
 from .api.tool_calling import (
-    build_json_system_prompt,
-    convert_tools_for_template,
-    extract_json_schema_for_guided,
-    parse_json_output,
-    parse_tool_calls,
+    build_json_system_prompt,  # noqa: F401
+    convert_tools_for_template,  # noqa: F401
+    extract_json_schema_for_guided,  # noqa: F401
+    parse_json_output,  # noqa: F401
+    parse_tool_calls,  # noqa: F401
 )
 from .api.utils import (
     SPECIAL_TOKENS_PATTERN,  # noqa: F401
-    StreamingThinkRouter,
-    StreamingToolCallFilter,
-    clean_output_text,
-    extract_json_from_response,
-    extract_multimodal_content,
+    StreamingThinkRouter,  # noqa: F401
+    StreamingToolCallFilter,  # noqa: F401
+    clean_output_text,  # noqa: F401
+    extract_json_from_response,  # noqa: F401
+    extract_multimodal_content,  # noqa: F401
     is_mllm_model,  # noqa: F401
-    sanitize_output,
-    strip_special_tokens,
-    strip_thinking_tags,
+    sanitize_output,  # noqa: F401
+    strip_special_tokens,  # noqa: F401
+    strip_thinking_tags,  # noqa: F401
 )
 from .config import get_config
 from .engine import (
     BaseEngine,
     BatchedEngine,
-    GenerationOutput,
     HybridEngine,
     SimpleEngine,
 )
 from .runtime.model_registry import ModelEntry, ModelRegistry
+from .service.helpers import (  # noqa: F401 — re-export for backward compat
+    _FALLBACK_TEMPERATURE,
+    _FALLBACK_TOP_P,
+    _TOOL_USE_SYSTEM_SUFFIX,
+    _build_usage,
+    _disconnect_guard,
+    _extract_token_logprob,
+    _inject_json_instruction,
+    _maybe_pin_system_prompt,
+    _parse_tool_calls_with_parser,
+    _resolve_max_tokens,
+    _resolve_model_name,
+    _resolve_temperature,
+    _resolve_top_p,
+    _validate_model_name,
+    _validate_tool_call_params,
+    _wait_with_disconnect,
+    get_engine,
+    get_usage,
+)
 from .tool_parsers import ToolParserManager
 
 logging.basicConfig(level=logging.INFO)
@@ -148,77 +163,10 @@ _default_timeout: float = 300.0  # Default request timeout in seconds (5 minutes
 _default_temperature: float | None = None  # Set via --default-temperature
 _default_top_p: float | None = None  # Set via --default-top-p
 
-_FALLBACK_TEMPERATURE = 0.7
-_FALLBACK_TOP_P = 0.9
 
-
-def _resolve_model_name(request_model: str | None) -> str:
-    """Resolve the model name for responses — never return literal 'default'."""
-    cfg = get_config()
-    if not request_model or request_model == "default":
-        return cfg.model_name or "default"
-    return request_model
-
-
-def _resolve_max_tokens(
-    request_value: int | None, enable_thinking: bool | None = None
-) -> int:
-    """Resolve max_tokens with thinking budget for reasoning models.
-
-    When a reasoning parser is active and the user requests a small max_tokens,
-    add extra headroom so <think>...</think> tokens don't eat into the content budget.
-    Skips the budget when the request explicitly disables thinking.
-    """
-    cfg = get_config()
-    base = request_value if request_value is not None else cfg.default_max_tokens
-    if enable_thinking is False:
-        return base  # User explicitly disabled thinking — no budget needed
-    if cfg.reasoning_parser_name and base > 0 and base < 4096:
-        return base + cfg.thinking_token_budget
-    return base
-
-
-def _build_usage(output: "GenerationOutput", reasoning_text: str | None) -> Usage:
-    """Build Usage with reasoning token breakdown when applicable."""
-    cfg = get_config()
-    total_completion = output.completion_tokens
-    if reasoning_text and cfg.reasoning_parser_name:
-        # Estimate reasoning tokens (~4 chars/token for English/Chinese mix)
-        reasoning_tokens = max(1, len(reasoning_text) // 4)
-        reasoning_tokens = min(reasoning_tokens, total_completion)
-        return Usage(
-            prompt_tokens=output.prompt_tokens,
-            completion_tokens=total_completion,
-            total_tokens=output.prompt_tokens + total_completion,
-            completion_tokens_details=CompletionTokensDetails(
-                reasoning_tokens=reasoning_tokens,
-            ),
-        )
-    return Usage(
-        prompt_tokens=output.prompt_tokens,
-        completion_tokens=total_completion,
-        total_tokens=output.prompt_tokens + total_completion,
-    )
-
-
-def _resolve_temperature(request_value: float | None) -> float:
-    """Resolve temperature: request > CLI default > fallback."""
-    if request_value is not None:
-        return request_value
-    cfg = get_config()
-    if cfg.default_temperature is not None:
-        return cfg.default_temperature
-    return _FALLBACK_TEMPERATURE
-
-
-def _resolve_top_p(request_value: float | None) -> float:
-    """Resolve top_p: request > CLI default > fallback."""
-    if request_value is not None:
-        return request_value
-    cfg = get_config()
-    if cfg.default_top_p is not None:
-        return cfg.default_top_p
-    return _FALLBACK_TOP_P
+# _FALLBACK_TEMPERATURE, _FALLBACK_TOP_P — moved to service/helpers.py
+# _resolve_model_name, _resolve_max_tokens, _build_usage,
+# _resolve_temperature, _resolve_top_p — moved to service/helpers.py
 
 
 # Global MCP manager
@@ -256,88 +204,10 @@ _no_thinking: bool = (
 _pin_system_prompt: bool = False  # Auto-pin system prompt prefix cache blocks
 _pinned_system_prompt_hash: str | None = None  # Hash of pinned system prompt
 
-# Tool-use system prompt (auto-injected when tools are provided and parser is active)
-_TOOL_USE_SYSTEM_SUFFIX = (
-    "\n\nIMPORTANT: When the user's request can be answered using the provided tools, "
-    "you MUST use the appropriate tool immediately. Do NOT ask for clarification when "
-    "a reasonable default exists. Do NOT explain what you will do — just do it. "
-    "Be direct and concise in your responses. "
-    "Do NOT think out loud or show your reasoning process. "
-    "Give direct answers only — no preamble like 'The user asks...' or 'Let me think...'."
-)
+# _TOOL_USE_SYSTEM_SUFFIX — moved to service/helpers.py
 
 
-def _maybe_pin_system_prompt(messages: list) -> None:
-    """Auto-pin system prompt prefix cache blocks on first request."""
-    cfg = get_config()
-
-    if not cfg.pin_system_prompt or cfg.engine is None:
-        return
-
-    # Extract system message content
-    system_content = None
-    for msg in messages:
-        role = msg.get("role") if isinstance(msg, dict) else getattr(msg, "role", None)
-        if role == "system":
-            content = (
-                msg.get("content")
-                if isinstance(msg, dict)
-                else getattr(msg, "content", None)
-            )
-            if isinstance(content, str):
-                system_content = content
-                break
-
-    if not system_content:
-        return
-
-    # Check if this system prompt is already pinned
-    prompt_hash = hashlib.sha256(system_content.encode()).hexdigest()[:16]
-    if prompt_hash == cfg.pinned_system_prompt_hash:
-        return  # Already pinned
-
-    # Try to pin via engine's cache manager
-    try:
-        # Get the tokenizer to encode the system prompt
-        tokenizer = None
-        if hasattr(cfg.engine, "_tokenizer"):
-            tokenizer = cfg.engine._tokenizer
-        elif hasattr(cfg.engine, "_model") and hasattr(cfg.engine._model, "tokenizer"):
-            tokenizer = cfg.engine._model.tokenizer
-
-        if tokenizer is None:
-            return
-
-        system_tokens = tokenizer.encode(system_content)
-        if not system_tokens or len(system_tokens) < 16:
-            return  # Too short to pin
-
-        # Try BlockAwarePrefixCache (paged cache mode)
-        if hasattr(cfg.engine, "_prefix_cache") and cfg.engine._prefix_cache is not None:
-            cache = cfg.engine._prefix_cache
-            if hasattr(cache, "pin_prefix"):
-                if cache.pin_prefix(system_tokens):
-                    cfg.pinned_system_prompt_hash = prompt_hash
-                    logger.info(
-                        f"Auto-pinned system prompt: {len(system_tokens)} tokens, "
-                        f"hash={prompt_hash}"
-                    )
-                    return
-
-        # Try PrefixCacheManager (trie-based cache mode)
-        if hasattr(cfg.engine, "_cache_manager") and cfg.engine._cache_manager is not None:
-            cache = cfg.engine._cache_manager
-            if hasattr(cache, "pin_prefix"):
-                if cache.pin_prefix(system_tokens):
-                    cfg.pinned_system_prompt_hash = prompt_hash
-                    logger.info(
-                        f"Auto-pinned system prompt (trie): {len(system_tokens)} tokens, "
-                        f"hash={prompt_hash}"
-                    )
-                    return
-
-    except Exception as e:
-        logger.debug(f"System prompt pinning failed: {e}")
+# _maybe_pin_system_prompt — moved to service/helpers.py
 
 
 def _load_prefix_cache_from_disk() -> None:
@@ -526,9 +396,9 @@ async def serialize_inference(request: Request, call_next):
 
 # Auth and rate limiting — moved to middleware/auth.py
 from .middleware.auth import (  # noqa: E402
-    RateLimiter,
-    check_rate_limit,
-    verify_api_key,
+    RateLimiter,  # noqa: F401
+    check_rate_limit,  # noqa: F401
+    verify_api_key,  # noqa: F401
 )
 from .middleware.auth import (
     rate_limiter as _rate_limiter,  # noqa: F401 — configured in main()
@@ -554,195 +424,10 @@ async def _global_exception_handler(request: Request, exc: Exception):
     )
 
 
-def get_engine(model_name: str | None = None) -> BaseEngine:
-    """Get the engine for a model, routing by name in multi-model mode.
-
-    Args:
-        model_name: Model name from request. None uses the default model.
-                    In single-model mode, this is ignored and the config
-                    engine is returned.
-    """
-    cfg = get_config()
-    # Multi-model path: registry has entries → route by name
-    if cfg.model_registry:
-        try:
-            return cfg.model_registry.get_engine(model_name)
-        except KeyError:
-            pass  # fall through to legacy path
-
-    # Single-model legacy path
-    if cfg.engine is None:
-        raise HTTPException(status_code=503, detail="Model not loaded")
-    return cfg.engine
+# get_engine, _validate_model_name — moved to service/helpers.py
 
 
-def _validate_model_name(request_model: str) -> None:
-    """Validate that the request model name matches a served model."""
-    if not request_model:
-        return
-
-    cfg = get_config()
-    # Multi-model: check registry
-    if cfg.model_registry and request_model in cfg.model_registry:
-        return
-    if cfg.model_registry and request_model == "default":
-        return
-
-    # Single-model: check config
-    if not cfg.model_name:
-        return
-    accepted = {cfg.model_name}
-    if cfg.model_alias:
-        accepted.add(cfg.model_alias)
-    if cfg.model_path:
-        accepted.add(cfg.model_path)
-    if request_model not in accepted:
-        available = (
-            ", ".join(cfg.model_registry.list_model_names())
-            if cfg.model_registry
-            else cfg.model_name
-        )
-        raise HTTPException(
-            status_code=404,
-            detail=f"The model `{request_model}` does not exist. "
-            f"Available: {available}",
-        )
-
-
-def _parse_tool_calls_with_parser(
-    output_text: str, request: ChatCompletionRequest | None = None
-) -> tuple[str, list | None]:
-    """
-    Parse tool calls from model output using the configured parser.
-
-    If --enable-auto-tool-choice is set with --tool-call-parser, uses the
-    selected parser. Otherwise falls back to the generic parse_tool_calls.
-
-    Args:
-        output_text: The model output text
-        request: The original request (for context)
-
-    Returns:
-        Tuple of (cleaned_text, tool_calls)
-    """
-    cfg = get_config()
-
-    request_dict = request.model_dump() if request else None
-
-    # If auto tool choice is not enabled, try inferred parser from reasoning parser
-    if not cfg.enable_auto_tool_choice or not cfg.tool_call_parser:
-        if cfg.reasoning_parser_name and request and request.tools:
-            _PARSER_MAP = {"minimax": "minimax"}
-            inferred = _PARSER_MAP.get(cfg.reasoning_parser_name)
-            if inferred:
-                try:
-                    parser_cls = ToolParserManager.get_tool_parser(inferred)
-                    tokenizer = getattr(cfg.engine, "_tokenizer", None)
-                    parser = parser_cls(tokenizer)
-                    parser.reset()
-                    result = parser.extract_tool_calls(output_text, request_dict)
-                    if result.tools_called:
-                        tool_calls = [
-                            ToolCall(
-                                id=tc.get("id", f"call_{uuid.uuid4().hex[:8]}"),
-                                type="function",
-                                function=FunctionCall(
-                                    name=tc["name"],
-                                    arguments=tc["arguments"],
-                                ),
-                            )
-                            for tc in result.tool_calls
-                        ]
-                        return result.content or "", tool_calls
-                except Exception as e:
-                    logger.debug(f"Auto-infer tool parser failed: {e}")
-        return parse_tool_calls(output_text, request_dict)
-
-    # Initialize parser if needed
-    if cfg.tool_parser_instance is None:
-        try:
-            parser_cls = ToolParserManager.get_tool_parser(cfg.tool_call_parser)
-            # Get tokenizer from engine if available
-            tokenizer = None
-            if cfg.engine is not None and hasattr(cfg.engine, "_tokenizer"):
-                tokenizer = cfg.engine._tokenizer
-            cfg.tool_parser_instance = parser_cls(tokenizer)
-            logger.info(f"Initialized tool call parser: {cfg.tool_call_parser}")
-        except Exception as e:
-            logger.warning(
-                f"Failed to initialize tool parser '{cfg.tool_call_parser}': {e}"
-            )
-            logger.warning("Falling back to generic parser")
-            return parse_tool_calls(output_text, request_dict)
-
-    # Use the configured parser
-    try:
-        # Reset parser state between requests
-        cfg.tool_parser_instance.reset()
-        result = cfg.tool_parser_instance.extract_tool_calls(output_text, request_dict)
-        if result.tools_called:
-            tool_calls = [
-                ToolCall(
-                    id=tc.get("id", f"call_{uuid.uuid4().hex[:8]}"),
-                    type="function",
-                    function=FunctionCall(
-                        name=tc["name"],
-                        arguments=tc["arguments"],
-                    ),
-                )
-                for tc in result.tool_calls
-            ]
-            return result.content or "", tool_calls
-        else:
-            # Fallback: specific parser didn't find tool calls,
-            # try generic parser which handles more formats (e.g. Nemotron XML)
-            return parse_tool_calls(output_text, request_dict)
-    except Exception as e:
-        logger.warning(f"Tool parser error: {e}")
-        return parse_tool_calls(output_text, request_dict)
-
-
-def _validate_tool_call_params(tool_calls: list, tools: list) -> None:
-    """
-    Validate tool call parameter values against their schemas (post-generation).
-
-    Logs warnings for invalid parameters but does not block the response.
-    This provides graceful degradation for SimpleEngine mode where logits-level
-    constraint is not available.
-    """
-    from .api.tool_logits import _extract_param_schemas, validate_param_value
-
-    tool_defs = [t.model_dump() if hasattr(t, "model_dump") else t for t in tools]
-    schemas = _extract_param_schemas(tool_defs)
-
-    for tc in tool_calls:
-        func = tc.function if hasattr(tc, "function") else tc.get("function", {})
-        func_name = func.name if hasattr(func, "name") else func.get("name", "")
-        args_str = (
-            func.arguments
-            if hasattr(func, "arguments")
-            else func.get("arguments", "{}")
-        )
-
-        try:
-            args = json.loads(args_str)
-        except (json.JSONDecodeError, ValueError):
-            logger.warning(
-                f"Tool call '{func_name}': arguments is not valid JSON: {args_str!r}"
-            )
-            continue
-
-        if not isinstance(args, dict):
-            continue
-
-        for param_name, param_value in args.items():
-            schema_key = f"{func_name}.{param_name}"
-            schema = schemas.get(schema_key)
-            if not schema:
-                continue
-            is_valid, error = validate_param_value(json.dumps(param_value), schema)
-            if not is_valid:
-                logger.warning(f"Tool call '{func_name}' param '{param_name}': {error}")
+# _parse_tool_calls_with_parser, _validate_tool_call_params — moved to service/helpers.py
 
 
 def _detect_native_tool_support() -> bool:
@@ -1043,72 +728,7 @@ def _sync_config() -> None:
     cfg.model_registry = _model_registry
 
 
-def _extract_token_logprob(
-    logprobs_array, token_id: int, tokenizer, top_k: int
-) -> TokenLogProb:
-    """
-    Convert an mx.array of log-probabilities to a TokenLogProb with top-k alternatives.
-
-    Args:
-        logprobs_array: mx.array of shape [vocab_size] with log-probabilities.
-        token_id: The actually sampled token ID.
-        tokenizer: Tokenizer for decoding token IDs to text.
-        top_k: Number of top alternatives to include.
-
-    Returns:
-        TokenLogProb with the sampled token and top-k alternatives.
-    """
-    import mlx.core as mx
-    import numpy as np
-
-    # Convert to float32 first — mx.array may be bfloat16 which numpy can't handle
-    if hasattr(logprobs_array, "astype"):
-        logprobs_array = logprobs_array.astype(mx.float32)
-    probs = np.array(logprobs_array).flatten()
-    # Get top-k indices
-    top_k = min(top_k, len(probs))
-    top_indices = np.argpartition(probs, -top_k)[-top_k:]
-    top_indices = top_indices[np.argsort(probs[top_indices])][::-1]
-
-    # Build top_logprobs list
-    top_logprobs = []
-    for idx in top_indices:
-        idx = int(idx)
-        tok_text = tokenizer.decode([idx])
-        tok_bytes = list(tok_text.encode("utf-8", errors="replace"))
-        top_logprobs.append(
-            TopLogProb(
-                token=tok_text,
-                logprob=float(probs[idx]),
-                bytes=tok_bytes,
-            )
-        )
-
-    # The sampled token
-    sampled_text = tokenizer.decode([token_id])
-    sampled_bytes = list(sampled_text.encode("utf-8", errors="replace"))
-
-    return TokenLogProb(
-        token=sampled_text,
-        logprob=float(probs[token_id]) if token_id < len(probs) else 0.0,
-        bytes=sampled_bytes,
-        top_logprobs=top_logprobs,
-    )
-
-
-def get_usage(output: GenerationOutput) -> Usage:
-    """Extract usage metrics from GenerationOutput."""
-    total_prompt_tokens = (
-        output.prompt_tokens if hasattr(output, "prompt_tokens") else 0
-    )
-    total_completion_tokens = (
-        output.completion_tokens if hasattr(output, "completion_tokens") else 0
-    )
-    return Usage(
-        prompt_tokens=total_prompt_tokens,
-        completion_tokens=total_completion_tokens,
-        total_tokens=total_prompt_tokens + total_completion_tokens,
-    )
+# _extract_token_logprob, get_usage — moved to service/helpers.py
 
 
 # =============================================================================
@@ -1124,1627 +744,21 @@ def get_usage(output: GenerationOutput) -> Usage:
 # =============================================================================
 
 
-# =============================================================================
-# Streaming disconnect detection
-# =============================================================================
+# _disconnect_guard, _wait_with_disconnect — moved to service/helpers.py
 
 
-async def _disconnect_guard(
-    generator: AsyncIterator[str],
-    raw_request: Request,
-    poll_interval: float = 0.5,
-) -> AsyncIterator[str]:
-    """Wrap streaming generator to abort on client disconnect.
+# Completion Endpoints — moved to routes/completions.py
 
-    Uses asyncio racing: each __anext__() on the inner generator is
-    raced against a disconnect poller.  This catches disconnects even
-    during prefill when no chunks are being yielded for tens of seconds.
 
-    On disconnect, aclose() propagates down the generator chain to
-    engine_core.stream_outputs() finally-block → abort_request().
-    """
-    import time as _time
+# Chat Completion Endpoints — moved to routes/chat.py
+# _inject_json_instruction — moved to service/helpers.py
 
-    _t0 = _time.monotonic()
 
-    def _elapsed():
-        return f"{_time.monotonic() - _t0:.1f}s"
+# Anthropic Messages API — moved to routes/anthropic.py
+# Re-export for backward compatibility (test_streaming_pipeline_integration)
+from .routes.anthropic import _emit_content_pieces  # noqa: F401, E402
 
-    logger.info(f"[disconnect_guard] START poll_interval={poll_interval}s")
-
-    async def _wait_disconnect():
-        poll_count = 0
-        while True:
-            await asyncio.sleep(poll_interval)
-            poll_count += 1
-            is_disc = await raw_request.is_disconnected()
-            if poll_count % 10 == 0 or is_disc:
-                logger.info(
-                    f"[disconnect_guard] poll #{poll_count} "
-                    f"disconnected={is_disc} elapsed={_elapsed()}"
-                )
-            if is_disc:
-                return
-
-    chunk_count = 0
-    disconnect_task: asyncio.Task | None = None
-    anext_task: asyncio.Task | None = None
-    try:
-        aiter = generator.__aiter__()
-        disconnect_task = asyncio.create_task(_wait_disconnect())
-        while True:
-            anext_task = asyncio.ensure_future(aiter.__anext__())
-            done, _ = await asyncio.wait(
-                [anext_task, disconnect_task],
-                return_when=asyncio.FIRST_COMPLETED,
-            )
-            if disconnect_task in done:
-                logger.info(
-                    f"[disconnect_guard] CLIENT DISCONNECTED after "
-                    f"{chunk_count} chunks, elapsed={_elapsed()}"
-                )
-                anext_task.cancel()
-                try:
-                    await anext_task
-                except (asyncio.CancelledError, StopAsyncIteration):
-                    pass
-                break
-            try:
-                chunk = anext_task.result()
-            except StopAsyncIteration:
-                logger.info(
-                    f"[disconnect_guard] generator exhausted normally, "
-                    f"{chunk_count} chunks, elapsed={_elapsed()}"
-                )
-                break
-            except Exception as exc:
-                logger.error(
-                    f"[disconnect_guard] generator raised {type(exc).__name__}: "
-                    f"{exc}, {chunk_count} chunks, elapsed={_elapsed()}",
-                    exc_info=True,
-                )
-                # Yield an error event so the client sees something before [DONE]
-                import json as _json
-
-                error_data = _json.dumps(
-                    {
-                        "error": {
-                            "message": f"Internal error during streaming: {exc}",
-                            "type": type(exc).__name__,
-                        }
-                    }
-                )
-                yield f"data: {error_data}\n\n"
-                yield "data: [DONE]\n\n"
-                break
-            chunk_count += 1
-            if chunk_count == 1:
-                logger.info(
-                    f"[disconnect_guard] first chunk arrived, elapsed={_elapsed()}"
-                )
-            yield chunk
-    except GeneratorExit:
-        logger.info(
-            f"[disconnect_guard] GeneratorExit after {chunk_count} chunks, elapsed={_elapsed()}"
-        )
-    finally:
-        if disconnect_task and not disconnect_task.done():
-            disconnect_task.cancel()
-        if anext_task and not anext_task.done():
-            anext_task.cancel()
-        # Close the generator to release resources (e.g., SimpleEngine's
-        # _generation_lock).  Without this, a client disconnect leaves the
-        # async generator suspended inside `async with _generation_lock`,
-        # blocking all subsequent requests on the single-user engine.
-        #
-        # For BatchedEngine with run_in_executor this was previously avoided
-        # due to potential Metal assertion failures, but task cancellation
-        # alone is insufficient for SimpleEngine's sync-loop generators.
-        try:
-            await generator.aclose()
-        except Exception:
-            pass
-        logger.info(
-            f"[disconnect_guard] CLEANUP done, {chunk_count} chunks total, elapsed={_elapsed()}"
-        )
-
-
-async def _wait_with_disconnect(
-    coro,
-    raw_request: Request,
-    timeout: float,
-    poll_interval: float = 0.5,
-):
-    """Run a coroutine with both timeout and client disconnect detection.
-
-    For non-streaming requests where _disconnect_guard() can't be used.
-    Races the coroutine against a disconnect poller, same pattern as
-    _disconnect_guard but for awaitable (non-generator) coroutines.
-    """
-    import time as _time
-
-    _t0 = _time.monotonic()
-
-    task = asyncio.ensure_future(coro)
-
-    async def _wait_disconnect():
-        poll_count = 0
-        while True:
-            await asyncio.sleep(poll_interval)
-            poll_count += 1
-            is_disc = await raw_request.is_disconnected()
-            if poll_count % 10 == 0 or is_disc:
-                logger.info(
-                    f"[disconnect_guard] poll #{poll_count} "
-                    f"disconnected={is_disc} elapsed={_time.monotonic() - _t0:.1f}s"
-                )
-            if is_disc:
-                return
-
-    disconnect_task = asyncio.create_task(_wait_disconnect())
-
-    try:
-        done, _ = await asyncio.wait(
-            [task, disconnect_task],
-            timeout=timeout,
-            return_when=asyncio.FIRST_COMPLETED,
-        )
-
-        if not done:
-            # Timeout
-            task.cancel()
-            try:
-                await task
-            except (asyncio.CancelledError, Exception):
-                pass
-            raise HTTPException(
-                status_code=504,
-                detail=f"Request timed out after {timeout:.1f} seconds",
-            )
-
-        if disconnect_task in done:
-            # Client disconnected
-            logger.info(
-                f"[disconnect_guard] CLIENT DISCONNECTED (non-stream) "
-                f"elapsed={_time.monotonic() - _t0:.1f}s"
-            )
-            task.cancel()
-            try:
-                await task
-            except (asyncio.CancelledError, Exception):
-                pass
-            return None  # Signal to caller that client disconnected
-
-        # Task completed
-        return task.result()
-
-    finally:
-        if not disconnect_task.done():
-            disconnect_task.cancel()
-        if not task.done():
-            task.cancel()
-
-
-# =============================================================================
-# Completion Endpoints
-# =============================================================================
-
-
-@app.post(
-    "/v1/completions", dependencies=[Depends(verify_api_key), Depends(check_rate_limit)]
-)
-async def create_completion(request: CompletionRequest, raw_request: Request):
-    """Create a text completion."""
-    _validate_model_name(request.model)
-    engine = get_engine(request.model)
-
-    # Handle single prompt or list of prompts
-    prompts = request.prompt if isinstance(request.prompt, list) else [request.prompt]
-
-    # --- Detailed request logging ---
-    prompt_preview = prompts[0][:200] if prompts else "(empty)"
-    prompt_len = sum(len(p) for p in prompts)
-    logger.info(
-        f"[REQUEST] POST /v1/completions stream={request.stream} "
-        f"max_tokens={request.max_tokens} temp={request.temperature} "
-        f"prompt_chars={prompt_len} prompt_preview={prompt_preview!r}"
-    )
-
-    if request.stream:
-        return StreamingResponse(
-            _disconnect_guard(
-                stream_completion(engine, prompts[0], request),
-                raw_request,
-            ),
-            media_type="text/event-stream",
-        )
-
-    # Non-streaming response with timing and timeout
-    start_time = time.perf_counter()
-    timeout = request.timeout or get_config().default_timeout
-    choices = []
-    total_completion_tokens = 0
-    total_prompt_tokens = 0
-
-    for i, prompt in enumerate(prompts):
-        output = await _wait_with_disconnect(
-            engine.generate(
-                prompt=prompt,
-                max_tokens=_resolve_max_tokens(request.max_tokens),
-                temperature=_resolve_temperature(request.temperature),
-                top_p=_resolve_top_p(request.top_p),
-                stop=request.stop,
-            ),
-            raw_request,
-            timeout=timeout,
-        )
-        if output is None:
-            return Response(status_code=499)  # Client closed request
-
-        choices.append(
-            CompletionChoice(
-                index=i,
-                text=output.text,
-                finish_reason=output.finish_reason,
-            )
-        )
-        total_completion_tokens += output.completion_tokens
-        total_prompt_tokens += (
-            output.prompt_tokens if hasattr(output, "prompt_tokens") else 0
-        )
-
-    elapsed = time.perf_counter() - start_time
-    tokens_per_sec = total_completion_tokens / elapsed if elapsed > 0 else 0
-    logger.info(
-        f"Completion: {total_prompt_tokens} prompt + {total_completion_tokens} completion tokens in {elapsed:.2f}s ({tokens_per_sec:.1f} tok/s)"
-    )
-
-    comp_response = CompletionResponse(
-        model=_resolve_model_name(request.model),
-        choices=choices,
-        usage=Usage(
-            prompt_tokens=total_prompt_tokens,
-            completion_tokens=total_completion_tokens,
-            total_tokens=total_prompt_tokens + total_completion_tokens,
-        ),
-    )
-    return Response(
-        content=comp_response.model_dump_json(exclude_none=True),
-        media_type="application/json",
-    )
-
-
-@app.post(
-    "/v1/chat/completions",
-    dependencies=[Depends(verify_api_key), Depends(check_rate_limit)],
-)
-async def create_chat_completion(request: ChatCompletionRequest, raw_request: Request):
-    """
-    Create a chat completion (supports multimodal content for VLM models).
-
-    OpenAI-compatible multimodal format for images:
-    ```json
-    messages=[{
-        "role": "user",
-        "content": [
-            {"type": "text", "text": "What's in this image?"},
-            {"type": "image_url", "image_url": {"url": "https://..."}}
-        ]
-    }]
-    ```
-
-    Video support:
-    ```json
-    messages=[{
-        "role": "user",
-        "content": [
-            {"type": "text", "text": "What happens in this video?"},
-            {"type": "video_url", "video_url": {"url": "https://example.com/video.mp4"}}
-        ]
-    }]
-    ```
-
-    Structured output (JSON mode):
-    ```json
-    response_format={"type": "json_object"}
-    ```
-
-    Structured output (JSON Schema):
-    ```json
-    response_format={
-        "type": "json_schema",
-        "json_schema": {
-            "name": "my_schema",
-            "schema": {"type": "object", "properties": {...}}
-        }
-    }
-    ```
-    """
-    _validate_model_name(request.model)
-    engine = get_engine(request.model)
-
-    # Validate messages is non-empty
-    if not request.messages:
-        raise HTTPException(
-            status_code=400,
-            detail="messages must not be empty",
-        )
-
-    # Validate model name matches loaded model
-    _validate_model_name(request.model)
-
-    # Validate message roles
-    _valid_roles = {"system", "user", "assistant", "tool", "developer"}
-    for msg in request.messages:
-        if msg.role not in _valid_roles:
-            raise HTTPException(
-                status_code=400,
-                detail=f"Invalid role '{msg.role}'. Must be one of: {', '.join(sorted(_valid_roles))}",
-            )
-
-    # Validate n parameter (only n=1 supported)
-    if request.n is not None and request.n > 1:
-        raise HTTPException(
-            status_code=400,
-            detail="n > 1 is not supported. Rapid-MLX generates one completion per request.",
-        )
-
-    # Validate max_tokens (must be positive)
-    if request.max_tokens is not None and request.max_tokens < 1:
-        raise HTTPException(
-            status_code=400,
-            detail="max_tokens must be at least 1",
-        )
-
-    # Validate temperature range (OpenAI spec: 0-2)
-    if request.temperature is not None and (
-        request.temperature < 0 or request.temperature > 2
-    ):
-        raise HTTPException(
-            status_code=400,
-            detail="temperature must be between 0 and 2",
-        )
-
-    # Validate top_logprobs range (OpenAI spec: 0-20)
-    if request.top_logprobs is not None and (
-        request.top_logprobs < 0 or request.top_logprobs > 20
-    ):
-        raise HTTPException(
-            status_code=400,
-            detail="top_logprobs must be between 0 and 20",
-        )
-
-    # --- Detailed request logging ---
-    n_msgs = len(request.messages)
-    msg_roles = [m.role for m in request.messages]
-    total_chars = 0
-    last_user_preview = ""
-    for m in request.messages:
-        content = m.content if isinstance(m.content, str) else str(m.content)
-        total_chars += len(content)
-        if m.role == "user":
-            last_user_preview = content[:300]
-    has_tools = bool(request.tools)
-    n_tools = len(request.tools) if request.tools else 0
-    logger.info(
-        f"[REQUEST] POST /v1/chat/completions stream={request.stream} "
-        f"model={request.model!r} max_tokens={request.max_tokens} "
-        f"temp={request.temperature} msgs={n_msgs} roles={msg_roles} "
-        f"total_chars={total_chars} tools={n_tools} "
-        f"response_format={request.response_format}"
-    )
-    logger.debug(f"[REQUEST] last user message preview: {last_user_preview!r}")
-
-    cfg = get_config()
-
-    # Save original messages (clean dicts) for cloud routing BEFORE
-    # local mutations (extract_multimodal_content, developer→system, suffix injection).
-    # Cloud APIs expect standard OpenAI-format messages.
-    if cfg.cloud_router:
-        _cloud_original_messages = [
-            msg.model_dump(exclude_none=True)
-            if hasattr(msg, "model_dump")
-            else {k: v for k, v in dict(msg).items() if v is not None}
-            for msg in request.messages
-        ]
-    else:
-        _cloud_original_messages = None
-
-    # For MLLM models, keep original messages with embedded images
-    # (MLLM.chat() extracts images from message content internally)
-    if engine.is_mllm:
-        # Convert Pydantic messages to dicts, excluding None fields
-        # to prevent chat templates from misinterpreting key presence
-        # (e.g. image_url: null on text parts triggers Qwen3-VL crash)
-        messages = []
-        for msg in request.messages:
-            if hasattr(msg, "model_dump"):
-                msg_dict = msg.model_dump(exclude_none=True)
-            else:
-                raw = dict(msg)
-                msg_dict = {k: v for k, v in raw.items() if v is not None}
-            messages.append(msg_dict)
-        images, videos = [], []  # MLLM extracts these from messages
-        logger.debug(f"MLLM: Processing {len(messages)} messages")
-    else:
-        # For LLM, extract text, images, and videos separately
-        messages, images, videos = extract_multimodal_content(
-            request.messages,
-            preserve_native_format=engine.preserve_native_tool_format,
-        )
-
-    has_media = bool(images or videos)
-    if engine.is_mllm and not has_media:
-        # MLLM extracts media from messages directly, so images/videos are
-        # always empty. Check message content for video/image types instead.
-        for msg in request.messages:
-            content = msg.content if hasattr(msg, "content") else msg.get("content", "")
-            if isinstance(content, list):
-                for item in content:
-                    item_type = (
-                        item.type
-                        if hasattr(item, "type")
-                        else (item.get("type", "") if isinstance(item, dict) else "")
-                    )
-                    if item_type in ("image_url", "image", "video", "video_url"):
-                        has_media = True
-                        break
-            if has_media:
-                break
-
-    # Normalize "developer" role to "system" for models that don't support it.
-    # OpenAI's API uses "developer" as a newer alias for "system", but most
-    # open-source chat templates (MiniMax, Qwen, Llama, etc.) only recognize
-    # "system". Without this, the entire system prompt gets silently dropped.
-    for i, m in enumerate(messages):
-        role = m.get("role") if isinstance(m, dict) else getattr(m, "role", None)
-        if role == "developer":
-            if isinstance(m, dict):
-                messages[i]["role"] = "system"
-            else:
-                m.role = "system"
-
-    # Auto-inject system prompt suffix for tool use and/or reasoning control
-    _inject_suffix = None
-    if request.tools and cfg.tool_call_parser:
-        _inject_suffix = _TOOL_USE_SYSTEM_SUFFIX
-    elif cfg.reasoning_parser_name == "minimax":
-        # MiniMax reasoning parser active — inject conciseness instruction
-        _inject_suffix = (
-            "\n\nDo NOT think out loud or show your reasoning process. "
-            "Give direct answers only — no preamble like 'The user asks...' or "
-            "'We should respond...' or 'Let me think...'. Be concise."
-        )
-
-    if _inject_suffix:
-        has_system = any(
-            (m.get("role") if isinstance(m, dict) else getattr(m, "role", None))
-            == "system"
-            for m in messages
-        )
-        if has_system:
-            for i, m in enumerate(messages):
-                role = (
-                    m.get("role") if isinstance(m, dict) else getattr(m, "role", None)
-                )
-                if role == "system":
-                    if isinstance(m, dict):
-                        messages[i] = {**m, "content": m["content"] + _inject_suffix}
-                    else:
-                        messages[i]["content"] = m["content"] + _inject_suffix
-                    break
-        else:
-            system_msg = {"role": "system", "content": _inject_suffix.strip()}
-            messages = [system_msg] + list(messages)
-
-    # Auto-pin system prompt prefix cache blocks (non-blocking, first request only)
-    if cfg.pin_system_prompt:
-        _maybe_pin_system_prompt(messages)
-
-    # Handle response_format - inject system prompt if needed
-    response_format = request.response_format
-    if response_format:
-        try:
-            json_instruction = build_json_system_prompt(response_format)
-        except Exception as e:
-            logger.warning(f"Failed to build JSON system prompt: {e}")
-            raise HTTPException(
-                status_code=400,
-                detail=f"Invalid response_format schema: {e}",
-            )
-        if json_instruction:
-            # Inject JSON instruction into messages
-            messages = _inject_json_instruction(messages, json_instruction)
-
-    # Prepare kwargs
-    chat_kwargs = {
-        "max_tokens": _resolve_max_tokens(request.max_tokens, request.enable_thinking),
-        "temperature": _resolve_temperature(request.temperature),
-        "top_p": _resolve_top_p(request.top_p),
-        "stop": request.stop,
-    }
-
-    # Add multimodal content
-    if has_media:
-        chat_kwargs["images"] = images if images else None
-        chat_kwargs["videos"] = videos if videos else None
-        if request.video_fps:
-            chat_kwargs["video_fps"] = request.video_fps
-        if request.video_max_frames:
-            chat_kwargs["video_max_frames"] = request.video_max_frames
-
-    # SpecPrefill: per-request overrides
-    if request.specprefill is not None:
-        chat_kwargs["specprefill"] = request.specprefill
-    if request.specprefill_keep_pct is not None:
-        chat_kwargs["specprefill_keep_pct"] = request.specprefill_keep_pct
-
-    # Add tools if provided
-    if request.tools:
-        chat_kwargs["tools"] = convert_tools_for_template(request.tools)
-
-    # Pass through enable_thinking if explicitly set by the client
-    if request.enable_thinking is not None:
-        chat_kwargs["enable_thinking"] = request.enable_thinking
-    elif cfg.no_thinking:
-        chat_kwargs["enable_thinking"] = False
-
-    # Cloud routing: offload large-context requests to cloud LLM
-    if cfg.cloud_router and not engine.is_mllm and hasattr(engine, "build_prompt"):
-        try:
-            prompt = engine.build_prompt(messages, tools=request.tools)
-            total_tokens, new_tokens = engine.model.estimate_new_tokens(prompt)
-            if cfg.cloud_router.should_route_to_cloud(new_tokens):
-                logger.info(
-                    f"[CLOUD ROUTE] {new_tokens} new tokens (total {total_tokens}) "
-                    f"> threshold {cfg.cloud_router.threshold}, "
-                    f"routing to {cfg.cloud_router.cloud_model}"
-                )
-                # Use pre-mutation messages for cloud (standard OpenAI format)
-                cloud_messages = _cloud_original_messages
-                cloud_kwargs = {
-                    "temperature": chat_kwargs.get("temperature"),
-                    "max_tokens": chat_kwargs.get("max_tokens"),
-                    "top_p": chat_kwargs.get("top_p"),
-                }
-                if request.stop:
-                    cloud_kwargs["stop"] = request.stop
-                if request.tool_choice is not None:
-                    cloud_kwargs["tool_choice"] = request.tool_choice
-                if request.response_format:
-                    rf = request.response_format
-                    cloud_kwargs["response_format"] = (
-                        rf.model_dump() if hasattr(rf, "model_dump") else rf
-                    )
-                if request.tools:
-                    # Pass raw tool defs (OpenAI format), not template-converted
-                    cloud_kwargs["tools"] = [
-                        t.model_dump() if hasattr(t, "model_dump") else t
-                        for t in request.tools
-                    ]
-                if request.stream:
-                    return StreamingResponse(
-                        _disconnect_guard(
-                            cfg.cloud_router.stream_completion(
-                                cloud_messages,
-                                model_name=cfg.model_name or "cloud",
-                                **cloud_kwargs,
-                            ),
-                            raw_request,
-                        ),
-                        media_type="text/event-stream",
-                    )
-                else:
-                    result = await _wait_with_disconnect(
-                        cfg.cloud_router.completion(cloud_messages, **cloud_kwargs),
-                        raw_request,
-                        timeout=request.timeout or cfg.default_timeout,
-                    )
-                    if result is None:
-                        # Client disconnected during cloud request
-                        return Response(status_code=499, content="Client disconnected")
-                    return Response(
-                        content=json.dumps(result),
-                        media_type="application/json",
-                    )
-            else:
-                logger.info(
-                    f"[LOCAL] {new_tokens} new tokens (total {total_tokens}) "
-                    f"<= threshold {cfg.cloud_router.threshold}, using local inference"
-                )
-        except Exception as e:
-            logger.warning(
-                f"[CLOUD ROUTE] Error during routing check: {e}, falling back to local"
-            )
-
-    if request.stream:
-        # Validate chat template eagerly so template errors return 400,
-        # not a broken SSE stream.  build_prompt is cheap (no generation).
-        if hasattr(engine, "build_prompt") and not engine.is_mllm:
-            try:
-                engine.build_prompt(
-                    messages,
-                    tools=chat_kwargs.get("tools"),
-                    enable_thinking=chat_kwargs.get("enable_thinking"),
-                )
-            except Exception as e:
-                err_msg = str(e)
-                err_type = type(e).__name__
-                if (
-                    "TemplateError" in err_type
-                    or "template" in err_msg.lower()
-                    or ("user" in err_msg.lower() and "found" in err_msg.lower())
-                ):
-                    raise HTTPException(
-                        status_code=400,
-                        detail=f"Chat template error: {err_msg}",
-                    )
-                raise
-        return StreamingResponse(
-            _disconnect_guard(
-                stream_chat_completion(engine, messages, request, **chat_kwargs),
-                raw_request,
-            ),
-            media_type="text/event-stream",
-        )
-
-    # Non-streaming response with timing and timeout
-    start_time = time.perf_counter()
-    timeout = request.timeout or cfg.default_timeout
-
-    # Disable GC during generation to avoid latency spikes
-    gc_was_enabled = gc.isenabled()
-    if cfg.gc_control and gc_was_enabled:
-        gc.disable()
-
-    # Determine if we need per-token logprobs
-    want_logprobs = request.logprobs and request.top_logprobs
-    top_k_logprobs = request.top_logprobs or 0
-    token_logprobs_list: list[TokenLogProb] = []
-
-    # Check if we should use guided generation for JSON schema
-    use_guided = False
-    json_schema = None
-    if response_format and not request.tools:
-        json_schema = extract_json_schema_for_guided(response_format)
-        if json_schema and hasattr(engine, "supports_guided_generation"):
-            use_guided = engine.supports_guided_generation
-            if use_guided:
-                logger.info("Using guided generation for JSON schema enforcement")
-
-    try:
-        if want_logprobs and not use_guided:
-            # Use streaming internally to collect per-token logprobs
-            output = None
-            async for chunk in engine.stream_chat(messages=messages, **chat_kwargs):
-                output = chunk
-                if chunk.logprobs is not None and chunk.new_text:
-                    token_id = chunk.tokens[-1] if chunk.tokens else 0
-                    token_logprobs_list.append(
-                        _extract_token_logprob(
-                            chunk.logprobs, token_id, engine.tokenizer, top_k_logprobs
-                        )
-                    )
-            if output is None:
-                return Response(status_code=499)
-        elif use_guided and json_schema:
-            # Use guided generation for constrained JSON output
-            # Fall back to standard generation if guided fails (bad schema, etc.)
-            try:
-                output = await _wait_with_disconnect(
-                    engine.generate_with_schema(
-                        messages=messages,
-                        json_schema=json_schema,
-                        **chat_kwargs,
-                    ),
-                    raw_request,
-                    timeout=timeout,
-                )
-            except Exception as guided_err:
-                logger.warning(
-                    f"Guided generation failed, falling back to standard: {guided_err}"
-                )
-                logger.debug(f"Problematic schema: {json_schema}")
-                output = await _wait_with_disconnect(
-                    engine.chat(messages=messages, **chat_kwargs),
-                    raw_request,
-                    timeout=timeout,
-                )
-        else:
-            # Standard generation
-            output = await _wait_with_disconnect(
-                engine.chat(messages=messages, **chat_kwargs),
-                raw_request,
-                timeout=timeout,
-            )
-    except HTTPException:
-        raise  # Don't wrap FastAPI HTTPExceptions
-    except Exception as e:
-        # Catch Jinja2 template errors (e.g. "No user query found in messages")
-        # and other chat template failures, return 400 instead of 500.
-        err_msg = str(e)
-        err_type = type(e).__name__
-        if (
-            "TemplateError" in err_type
-            or "template" in err_msg.lower()
-            or ("user" in err_msg.lower() and "found" in err_msg.lower())
-        ):
-            raise HTTPException(
-                status_code=400, detail=f"Chat template error: {err_msg}"
-            )
-        raise  # Re-raise non-template errors
-    finally:
-        if cfg.gc_control and gc_was_enabled:
-            gc.enable()
-            gc.collect()
-
-    if output is None:
-        return Response(status_code=499)  # Client closed request
-
-    elapsed = time.perf_counter() - start_time
-    tokens_per_sec = output.completion_tokens / elapsed if elapsed > 0 else 0
-    logger.info(
-        f"Chat completion: {output.completion_tokens} tokens in {elapsed:.2f}s ({tokens_per_sec:.1f} tok/s)"
-    )
-
-    # Parse tool calls from output using configured parser
-    cleaned_text, tool_calls = _parse_tool_calls_with_parser(output.text, request)
-
-    # Validate tool call parameter values against schemas (SimpleEngine post-generation check)
-    if tool_calls and request.tools:
-        _validate_tool_call_params(tool_calls, request.tools)
-
-    # Extract reasoning content FIRST (strips channel tokens before JSON extraction)
-    # Always extract reasoning regardless of tool_calls — reasoning preamble
-    # should never leak into content even when tool calls are present.
-    reasoning_text = None
-    if cfg.reasoning_parser:
-        text_to_parse = cleaned_text or output.text
-        reasoning_text, cleaned_text = cfg.reasoning_parser.extract_reasoning(
-            text_to_parse
-        )
-
-    # Process response_format if specified (after reasoning parser cleaned the text)
-    if response_format and not tool_calls:
-        json_input = cleaned_text or output.text
-        try:
-            _, parsed_json, is_valid, error = parse_json_output(
-                json_input, response_format
-            )
-            if parsed_json is not None:
-                # Return JSON as string
-                cleaned_text = json.dumps(parsed_json)
-            if not is_valid:
-                logger.warning(f"JSON validation failed: {error}")
-        except Exception as e:
-            logger.warning(f"JSON output parsing failed: {e}")
-            # Don't crash — return the raw text as-is
-
-    # Determine finish reason
-    finish_reason = "tool_calls" if tool_calls else output.finish_reason
-
-    # Clean and strip thinking tags from content
-    # Thinking tags break JSON parsing in clients expecting pure content
-    # Also extract JSON if model outputs reasoning text before JSON
-    final_content = None
-    if cleaned_text:
-        final_content = strip_thinking_tags(clean_output_text(cleaned_text))
-        # Final catch-all sanitizer — strips any leftover tool-call /
-        # channel markup that escaped the parsers (e.g. Gemma 4 token-level
-        # routing artifacts when stream=False bypasses the streaming filters).
-        final_content = sanitize_output(final_content)
-        # If JSON mode requested, extract JSON from reasoning text
-        # (e.g., Qwen3 reasoning mode: "Let me think... {json}")
-        if response_format and final_content:
-            final_content = extract_json_from_response(final_content)
-
-    # Build logprobs for response if requested
-    choice_logprobs = None
-    if want_logprobs and token_logprobs_list:
-        choice_logprobs = ChoiceLogProbs(content=token_logprobs_list)
-
-    chat_response = ChatCompletionResponse(
-        model=_resolve_model_name(request.model),
-        choices=[
-            ChatCompletionChoice(
-                message=AssistantMessage(
-                    content=final_content,
-                    reasoning_content=reasoning_text,
-                    tool_calls=tool_calls,
-                ),
-                finish_reason=finish_reason,
-                logprobs=choice_logprobs,
-            )
-        ],
-        usage=_build_usage(output, reasoning_text),
-    )
-    return Response(
-        content=chat_response.model_dump_json(exclude_none=True),
-        media_type="application/json",
-    )
-
-
-def _inject_json_instruction(messages: list, instruction: str) -> list:
-    """
-    Inject JSON instruction into messages.
-
-    PREPENDS instruction to system message for better instruction following.
-    JSON formatting instructions at the beginning of system message are more
-    likely to be followed than instructions at the end.
-
-    If a system message exists, prepend to it. Otherwise, prepend a new system message.
-    """
-    messages = list(messages)  # Make a copy
-
-    # Find existing system message
-    system_idx = None
-    for i, msg in enumerate(messages):
-        role = msg.get("role") if isinstance(msg, dict) else getattr(msg, "role", None)
-        if role == "system":
-            system_idx = i
-            break
-
-    if system_idx is not None:
-        # PREPEND to existing system message (not append!)
-        # Instructions at the start are more effective
-        msg = messages[system_idx]
-        if isinstance(msg, dict):
-            existing = msg.get("content", "")
-            msg["content"] = f"{instruction}\n\n{existing}"
-        else:
-            existing = getattr(msg, "content", "") or ""
-            msg.content = f"{instruction}\n\n{existing}"
-    else:
-        # Prepend new system message
-        messages.insert(0, {"role": "system", "content": instruction})
-
-    return messages
-
-
-# =============================================================================
-# Anthropic Messages API Endpoints
-# =============================================================================
-
-
-@app.post("/v1/messages")
-async def create_anthropic_message(
-    request: Request,
-):
-    """
-    Anthropic Messages API endpoint.
-
-    Translates Anthropic-format requests to OpenAI format, runs inference
-    through the existing engine, and converts the response back.
-
-    Supports both streaming and non-streaming modes.
-    """
-    # Parse the raw body to handle Anthropic request format
-    body = await request.json()
-    anthropic_request = AnthropicRequest(**body)
-
-    _validate_model_name(anthropic_request.model)
-    engine = get_engine(anthropic_request.model)
-
-    # --- Detailed request logging ---
-    n_msgs = len(anthropic_request.messages)
-    total_chars = 0
-    last_user_preview = ""
-    for m in anthropic_request.messages:
-        content = m.content if isinstance(m.content, str) else str(m.content)
-        total_chars += len(content)
-        if m.role == "user":
-            last_user_preview = content[:300]
-    sys_chars = len(anthropic_request.system) if anthropic_request.system else 0
-    n_tools = len(anthropic_request.tools) if anthropic_request.tools else 0
-    logger.info(
-        f"[REQUEST] POST /v1/messages (anthropic) stream={anthropic_request.stream} "
-        f"model={anthropic_request.model!r} max_tokens={anthropic_request.max_tokens} "
-        f"msgs={n_msgs} total_chars={total_chars} system_chars={sys_chars} "
-        f"tools={n_tools}"
-    )
-    logger.debug(f"[REQUEST] last user message preview: {last_user_preview!r}")
-
-    # Validate model name matches loaded model
-    _validate_model_name(anthropic_request.model)
-
-    # Convert Anthropic request -> OpenAI request
-    openai_request = anthropic_to_openai(anthropic_request)
-
-    if anthropic_request.stream:
-        return StreamingResponse(
-            _disconnect_guard(
-                _stream_anthropic_messages(engine, openai_request, anthropic_request),
-                request,
-            ),
-            media_type="text/event-stream",
-            headers={
-                "Cache-Control": "no-cache",
-                "Connection": "keep-alive",
-            },
-        )
-
-    # Non-streaming: run inference through existing engine
-    messages, images, videos = extract_multimodal_content(
-        openai_request.messages,
-        preserve_native_format=engine.preserve_native_tool_format,
-    )
-
-    chat_kwargs = {
-        "max_tokens": _resolve_max_tokens(
-            openai_request.max_tokens,
-            getattr(openai_request, "enable_thinking", None),
-        ),
-        "temperature": openai_request.temperature,
-        "top_p": openai_request.top_p,
-    }
-
-    if openai_request.tools:
-        chat_kwargs["tools"] = convert_tools_for_template(openai_request.tools)
-    cfg = get_config()
-    if openai_request.enable_thinking is not None:
-        chat_kwargs["enable_thinking"] = openai_request.enable_thinking
-    elif cfg.no_thinking:
-        chat_kwargs["enable_thinking"] = False
-
-    start_time = time.perf_counter()
-    timeout = cfg.default_timeout
-
-    try:
-        output = await _wait_with_disconnect(
-            engine.chat(messages=messages, **chat_kwargs),
-            request,
-            timeout=timeout,
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        err_msg = str(e)
-        err_type = type(e).__name__
-        if (
-            "TemplateError" in err_type
-            or "template" in err_msg.lower()
-            or ("user" in err_msg.lower() and "found" in err_msg.lower())
-        ):
-            raise HTTPException(
-                status_code=400, detail=f"Chat template error: {err_msg}"
-            )
-        raise
-    if output is None:
-        return Response(status_code=499)  # Client closed request
-
-    elapsed = time.perf_counter() - start_time
-    tokens_per_sec = output.completion_tokens / elapsed if elapsed > 0 else 0
-    logger.info(
-        f"Anthropic messages: {output.completion_tokens} tokens in {elapsed:.2f}s ({tokens_per_sec:.1f} tok/s)"
-    )
-
-    # Parse tool calls
-    cleaned_text, tool_calls = _parse_tool_calls_with_parser(
-        output.text, openai_request
-    )
-
-    # Clean output text — strip think tags so Anthropic clients get pure content
-    final_content = None
-    if cleaned_text:
-        final_content = strip_thinking_tags(clean_output_text(cleaned_text))
-
-    # Determine finish reason
-    finish_reason = "tool_calls" if tool_calls else output.finish_reason
-
-    # Build OpenAI response to convert
-    openai_response = ChatCompletionResponse(
-        model=cfg.model_name or openai_request.model,
-        choices=[
-            ChatCompletionChoice(
-                message=AssistantMessage(
-                    content=final_content,
-                    tool_calls=tool_calls,
-                ),
-                finish_reason=finish_reason,
-            )
-        ],
-        usage=_build_usage(output, None),
-    )
-
-    # Convert to Anthropic response
-    anthropic_response = openai_to_anthropic(
-        openai_response, cfg.model_name or anthropic_request.model
-    )
-    return Response(
-        content=anthropic_response.model_dump_json(exclude_none=True),
-        media_type="application/json",
-    )
-
-
-@app.post("/v1/messages/count_tokens")
-async def count_anthropic_tokens(request: Request):
-    """
-    Count tokens for an Anthropic Messages API request.
-
-    Uses the model's tokenizer for accurate counting.
-    Claude Code calls this endpoint for token budgeting.
-    Note: Don't parse via AnthropicRequest — count_tokens requests
-    from Claude Code don't include max_tokens.
-    """
-    body = await request.json()
-
-    engine = get_engine()
-    tokenizer = engine.tokenizer
-
-    total_tokens = 0
-
-    # System message
-    system = body.get("system", "")
-    if isinstance(system, str) and system:
-        total_tokens += len(tokenizer.encode(system))
-    elif isinstance(system, list):
-        for block in system:
-            if isinstance(block, dict):
-                text = block.get("text", "")
-                if text:
-                    total_tokens += len(tokenizer.encode(text))
-
-    # Messages
-    for msg in body.get("messages", []):
-        content = msg.get("content", "")
-        if isinstance(content, str):
-            if content:
-                total_tokens += len(tokenizer.encode(content))
-        elif isinstance(content, list):
-            for block in content:
-                if isinstance(block, dict):
-                    text = block.get("text", "")
-                    if text:
-                        total_tokens += len(tokenizer.encode(text))
-                    # tool_use input
-                    if block.get("input"):
-                        total_tokens += len(
-                            tokenizer.encode(json.dumps(block["input"]))
-                        )
-                    # tool_result content
-                    sub_content = block.get("content", "")
-                    if isinstance(sub_content, str) and sub_content:
-                        total_tokens += len(tokenizer.encode(sub_content))
-                    elif isinstance(sub_content, list):
-                        for item in sub_content:
-                            if isinstance(item, dict):
-                                item_text = item.get("text", "")
-                                if item_text:
-                                    total_tokens += len(tokenizer.encode(item_text))
-
-    # Tools
-    for tool in body.get("tools", []):
-        name = tool.get("name", "")
-        if name:
-            total_tokens += len(tokenizer.encode(name))
-        desc = tool.get("description", "")
-        if desc:
-            total_tokens += len(tokenizer.encode(desc))
-        if tool.get("input_schema"):
-            total_tokens += len(tokenizer.encode(json.dumps(tool["input_schema"])))
-
-    return {"input_tokens": total_tokens}
-
-
-def _emit_content_pieces(
-    pieces: list[tuple[str, str]],
-    current_block_type: str | None,
-    block_index: int,
-) -> tuple[list[str], str | None, int]:
-    """Emit Anthropic SSE events for content pieces from the think router.
-
-    Handles block type transitions (thinking <-> text), emitting
-    content_block_start/stop/delta events as needed.
-
-    Args:
-        pieces: List of (block_type, text) from StreamingThinkRouter
-        current_block_type: Current open block type, or None
-        block_index: Current block index
-
-    Returns:
-        Tuple of (events, updated_block_type, updated_block_index)
-    """
-    events = []
-    for block_type, text in pieces:
-        if block_type != current_block_type:
-            # Close previous block if open
-            if current_block_type is not None:
-                events.append(
-                    f"event: content_block_stop\ndata: "
-                    f"{json.dumps({'type': 'content_block_stop', 'index': block_index})}\n\n"
-                )
-                block_index += 1
-            # Start new block
-            current_block_type = block_type
-            content_block = (
-                {"type": block_type, "text": ""}
-                if block_type == "text"
-                else {"type": block_type, "thinking": ""}
-            )
-            events.append(
-                f"event: content_block_start\ndata: "
-                f"{json.dumps({'type': 'content_block_start', 'index': block_index, 'content_block': content_block})}\n\n"
-            )
-        # Emit delta
-        delta_key = "thinking" if block_type == "thinking" else "text"
-        delta_type = "thinking_delta" if block_type == "thinking" else "text_delta"
-        delta_event = {
-            "type": "content_block_delta",
-            "index": block_index,
-            "delta": {"type": delta_type, delta_key: text},
-        }
-        events.append(
-            f"event: content_block_delta\ndata: {json.dumps(delta_event)}\n\n"
-        )
-    return events, current_block_type, block_index
-
-
-async def _stream_anthropic_messages(
-    engine: BaseEngine,
-    openai_request: ChatCompletionRequest,
-    anthropic_request: AnthropicRequest,
-) -> AsyncIterator[str]:
-    """
-    Stream Anthropic Messages API SSE events.
-
-    Converts OpenAI streaming chunks to Anthropic event format:
-    message_start -> content_block_start -> content_block_delta* ->
-    content_block_stop -> message_delta -> message_stop
-    """
-    msg_id = f"msg_{uuid.uuid4().hex[:24]}"
-    start_time = time.perf_counter()
-
-    # Extract messages for engine
-    messages, images, videos = extract_multimodal_content(
-        openai_request.messages,
-        preserve_native_format=engine.preserve_native_tool_format,
-    )
-
-    chat_kwargs = {
-        "max_tokens": _resolve_max_tokens(
-            openai_request.max_tokens,
-            getattr(openai_request, "enable_thinking", None),
-        ),
-        "temperature": openai_request.temperature,
-        "top_p": openai_request.top_p,
-    }
-
-    if openai_request.tools:
-        chat_kwargs["tools"] = convert_tools_for_template(openai_request.tools)
-    cfg = get_config()
-    if openai_request.enable_thinking is not None:
-        chat_kwargs["enable_thinking"] = openai_request.enable_thinking
-    elif cfg.no_thinking:
-        chat_kwargs["enable_thinking"] = False
-
-    # Emit message_start
-    message_start = {
-        "type": "message_start",
-        "message": {
-            "id": msg_id,
-            "type": "message",
-            "role": "assistant",
-            "model": cfg.model_name or anthropic_request.model,
-            "content": [],
-            "stop_reason": None,
-            "stop_sequence": None,
-            "usage": {
-                "input_tokens": 0,
-                "output_tokens": 0,
-            },
-        },
-    }
-    yield f"event: message_start\ndata: {json.dumps(message_start)}\n\n"
-
-    # Stream pipeline: raw text → tool call filter → think router → emit
-    # - Tool call filter strips tool call markup (emitted as structured blocks later)
-    # - Think router separates <think> content into Anthropic thinking blocks
-    accumulated_text = ""
-    accumulated_raw = ""
-    tool_filter = StreamingToolCallFilter()
-    # Detect if the model's chat template injects <think> into the
-    # generation prompt. If so, the model starts in thinking mode and
-    # the opening tag never appears in the output stream.
-    _tokenizer = engine.tokenizer if hasattr(engine, "tokenizer") else None
-    _chat_template = ""
-    if _tokenizer and hasattr(_tokenizer, "chat_template"):
-        _chat_template = _tokenizer.chat_template or ""
-    _starts_thinking = (
-        "<think>" in _chat_template and "add_generation_prompt" in _chat_template
-    )
-    think_router = StreamingThinkRouter(start_in_thinking=_starts_thinking)
-    prompt_tokens = 0
-    completion_tokens = 0
-
-    # Track which content blocks we've started
-    current_block_type = None  # "thinking" or "text"
-    block_index = 0
-
-    # Reset reasoning parser state for this stream
-    if cfg.reasoning_parser:
-        cfg.reasoning_parser.reset_state()
-
-    async for output in engine.stream_chat(messages=messages, **chat_kwargs):
-        delta_text = output.new_text
-
-        # Track token counts
-        if hasattr(output, "prompt_tokens") and output.prompt_tokens:
-            prompt_tokens = output.prompt_tokens
-        if hasattr(output, "completion_tokens") and output.completion_tokens:
-            completion_tokens = output.completion_tokens
-
-        if delta_text:
-            # Accumulate raw text BEFORE special token cleaning for tool parsing
-            accumulated_text += delta_text
-            content = None
-
-            # Use reasoning parser to separate reasoning from content
-            if cfg.reasoning_parser:
-                previous_raw = accumulated_raw
-                accumulated_raw += delta_text
-                delta_msg = cfg.reasoning_parser.extract_reasoning_streaming(
-                    previous_raw, accumulated_raw, delta_text
-                )
-                if delta_msg is not None:
-                    # Only emit content, discard reasoning for Anthropic clients
-                    content = delta_msg.content
-            else:
-                # No reasoning parser — pass through with special token filter
-                content = strip_special_tokens(delta_text)
-
-            if content:
-                # Filter special tokens from parser output too
-                content = strip_special_tokens(content)
-
-            if content:
-                # Stage 1: strip tool call markup
-                filtered = tool_filter.process(content)
-                if not filtered:
-                    continue
-                # Stage 2: route thinking vs text
-                pieces = think_router.process(filtered)
-                events, current_block_type, block_index = _emit_content_pieces(
-                    pieces, current_block_type, block_index
-                )
-                for event in events:
-                    yield event
-
-    # Flush remaining from both filters
-    remaining = tool_filter.flush()
-    if remaining:
-        events, current_block_type, block_index = _emit_content_pieces(
-            think_router.process(remaining), current_block_type, block_index
-        )
-        for event in events:
-            yield event
-
-    flush_pieces = think_router.flush()
-    if flush_pieces:
-        events, current_block_type, block_index = _emit_content_pieces(
-            flush_pieces, current_block_type, block_index
-        )
-        for event in events:
-            yield event
-
-    # Close final content block
-    if current_block_type is not None:
-        yield f"event: content_block_stop\ndata: {json.dumps({'type': 'content_block_stop', 'index': block_index})}\n\n"
-        block_index += 1
-
-    # Handle reasoning parser finalization (e.g. no-tag correction)
-    if cfg.reasoning_parser and accumulated_raw:
-        final_msg = (
-            cfg.reasoning_parser.finalize_streaming(accumulated_raw)
-            if hasattr(cfg.reasoning_parser, "finalize_streaming")
-            else None
-        )
-        if final_msg and final_msg.content:
-            # Emit corrected content (model didn't use think tags at all).
-            # At this point all previous blocks are closed (line above),
-            # so we must open a new text block before sending the delta.
-            content = strip_special_tokens(final_msg.content)
-            if content:
-                accumulated_text = content  # Replace accumulated
-                # Open a text block at current block_index
-                yield (
-                    f"event: content_block_start\ndata: "
-                    f"{json.dumps({'type': 'content_block_start', 'index': block_index, 'content_block': {'type': 'text', 'text': ''}})}\n\n"
-                )
-                delta_event = {
-                    "type": "content_block_delta",
-                    "index": block_index,
-                    "delta": {"type": "text_delta", "text": content},
-                }
-                yield f"event: content_block_delta\ndata: {json.dumps(delta_event)}\n\n"
-                # Close the block
-                yield f"event: content_block_stop\ndata: {json.dumps({'type': 'content_block_stop', 'index': block_index})}\n\n"
-                block_index += 1
-        # Reset parser state for next request
-        cfg.reasoning_parser.reset_state()
-
-    # Check for tool calls in accumulated text
-    _, tool_calls = _parse_tool_calls_with_parser(accumulated_text, openai_request)
-
-    # If there are tool calls, emit tool_use blocks
-    if tool_calls:
-        for i, tc in enumerate(tool_calls):
-            tool_index = block_index + i
-            try:
-                tool_input = json.loads(tc.function.arguments)
-            except (json.JSONDecodeError, AttributeError):
-                tool_input = {}
-
-            # content_block_start for tool_use
-            tool_block_start = {
-                "type": "content_block_start",
-                "index": tool_index,
-                "content_block": {
-                    "type": "tool_use",
-                    "id": tc.id,
-                    "name": tc.function.name,
-                    "input": {},
-                },
-            }
-            yield f"event: content_block_start\ndata: {json.dumps(tool_block_start)}\n\n"
-
-            # Send input as a single delta
-            input_json = json.dumps(tool_input)
-            input_delta = {
-                "type": "content_block_delta",
-                "index": tool_index,
-                "delta": {"type": "input_json_delta", "partial_json": input_json},
-            }
-            yield f"event: content_block_delta\ndata: {json.dumps(input_delta)}\n\n"
-
-            # content_block_stop
-            yield f"event: content_block_stop\ndata: {json.dumps({'type': 'content_block_stop', 'index': tool_index})}\n\n"
-
-    # Determine stop reason
-    stop_reason = "tool_use" if tool_calls else "end_turn"
-
-    # Emit message_delta with stop_reason and usage
-    message_delta = {
-        "type": "message_delta",
-        "delta": {"stop_reason": stop_reason, "stop_sequence": None},
-        "usage": {"input_tokens": prompt_tokens, "output_tokens": completion_tokens},
-    }
-    yield f"event: message_delta\ndata: {json.dumps(message_delta)}\n\n"
-
-    # Log throughput
-    elapsed = time.perf_counter() - start_time
-    tokens_per_sec = completion_tokens / elapsed if elapsed > 0 else 0
-    logger.info(
-        f"Anthropic messages (stream): prompt={prompt_tokens} + completion={completion_tokens} tokens in {elapsed:.2f}s ({tokens_per_sec:.1f} tok/s)"
-    )
-
-    # Emit message_stop
-    yield f"event: message_stop\ndata: {json.dumps({'type': 'message_stop'})}\n\n"
-
-
-# =============================================================================
-# Streaming Helpers
-# =============================================================================
-
-
-async def stream_completion(
-    engine: BaseEngine,
-    prompt: str,
-    request: CompletionRequest,
-) -> AsyncIterator[str]:
-    """Stream completion response."""
-    async for output in engine.stream_generate(
-        prompt=prompt,
-        max_tokens=_resolve_max_tokens(request.max_tokens),
-        temperature=_resolve_temperature(request.temperature),
-        top_p=_resolve_top_p(request.top_p),
-        stop=request.stop,
-    ):
-        data = {
-            "id": f"cmpl-{uuid.uuid4().hex[:8]}",
-            "object": "text_completion",
-            "created": int(time.time()),
-            "model": _resolve_model_name(request.model),
-            "choices": [
-                {
-                    "index": 0,
-                    "text": output.new_text,
-                    "finish_reason": output.finish_reason if output.finished else None,
-                }
-            ],
-        }
-        if output.finished:
-            data["usage"] = get_usage(output).model_dump()
-        yield f"data: {json.dumps(data)}\n\n"
-
-    yield "data: [DONE]\n\n"
-
-
-async def stream_chat_completion(
-    engine: BaseEngine,
-    messages: list,
-    request: ChatCompletionRequest,
-    **kwargs,
-) -> AsyncIterator[str]:
-    """Stream chat completion response.
-
-    Uses StreamingPostProcessor for reasoning/tool/sanitization pipeline.
-    SSE formatting stays inline for performance (fast path bypasses Pydantic).
-    """
-    from .service.postprocessor import StreamingPostProcessor
-
-    cfg = get_config()
-    gc_was_enabled = gc.isenabled()
-    if cfg.gc_control and gc_was_enabled:
-        gc.disable()
-
-    try:
-        response_id = f"chatcmpl-{uuid.uuid4().hex[:8]}"
-        start_time = time.perf_counter()
-
-        # Check if we should include usage in the final chunk
-        include_usage = request.stream_options and request.stream_options.include_usage
-
-        # Logprobs configuration
-        want_logprobs = request.logprobs and request.top_logprobs
-        top_k_logprobs = request.top_logprobs or 0
-
-        def _build_chunk_logprobs(output: GenerationOutput) -> ChoiceLogProbs | None:
-            """Build ChoiceLogProbs for a streaming chunk if logprobs requested."""
-            if not want_logprobs or output.logprobs is None:
-                return None
-            token_id = output.tokens[-1] if output.tokens else 0
-            token_lp = _extract_token_logprob(
-                output.logprobs, token_id, engine.tokenizer, top_k_logprobs
-            )
-            return ChoiceLogProbs(content=[token_lp])
-
-        # Pre-compute SSE template parts that don't change per-token.
-        _sse_created = int(time.time())
-        _model_escaped = json.dumps(_resolve_model_name(request.model))
-        _sse_prefix = (
-            f'data: {{"id":"{response_id}","object":"chat.completion.chunk",'
-            f'"created":{_sse_created},"model":{_model_escaped},'
-            f'"choices":[{{"index":0,"delta":{{'
-        )
-        _sse_suffix = "}}]}\n\n"
-
-        def _fast_sse_chunk(text: str, field: str = "content") -> str:
-            """Build SSE chunk JSON directly, bypassing Pydantic serialization."""
-            escaped = json.dumps(text)
-            return f'{_sse_prefix}"{field}":{escaped}{_sse_suffix}'
-
-        # First chunk with role
-        _first_sse = f'{_sse_prefix}"role":"assistant"{_sse_suffix}'
-        if logger.isEnabledFor(logging.INFO):
-            logger.info(f"[SSE-ROLE] {_first_sse.strip()[:200]}")
-        yield _first_sse
-
-        # Initialize post-processor (replaces ~40 lines of parser init)
-        processor = StreamingPostProcessor(
-            cfg, tools_requested=bool(request.tools)
-        )
-        processor.set_thinking_model(request.model)
-        processor.reset()
-
-        # Track token counts for usage reporting
-        prompt_tokens = 0
-        completion_tokens = 0
-
-        # Stream content — PostProcessor handles reasoning/tool/sanitize
-        async for output in engine.stream_chat(messages=messages, **kwargs):
-            # Track token counts from output (updated each chunk)
-            if hasattr(output, "prompt_tokens") and output.prompt_tokens:
-                prompt_tokens = output.prompt_tokens
-            if hasattr(output, "completion_tokens") and output.completion_tokens:
-                completion_tokens = output.completion_tokens
-
-            for event in processor.process_chunk(output):
-                if event.type == "content":
-                    # Fast path: simple content delta — skip Pydantic
-                    if not want_logprobs:
-                        _sse = _fast_sse_chunk(event.content, "content")
-                        if _sse:
-                            yield _sse
-                    else:
-                        # Pydantic path: need logprobs attached
-                        chunk = ChatCompletionChunk(
-                            id=response_id,
-                            model=_resolve_model_name(request.model),
-                            choices=[
-                                ChatCompletionChunkChoice(
-                                    delta=ChatCompletionChunkDelta(
-                                        content=event.content,
-                                    ),
-                                    logprobs=_build_chunk_logprobs(output),
-                                )
-                            ],
-                        )
-                        yield f"data: {chunk.model_dump_json(exclude_none=True)}\n\n"
-
-                elif event.type == "reasoning":
-                    yield _fast_sse_chunk(event.reasoning, "reasoning_content")
-
-                elif event.type == "tool_call":
-                    chunk = ChatCompletionChunk(
-                        id=response_id,
-                        model=_resolve_model_name(request.model),
-                        choices=[
-                            ChatCompletionChunkChoice(
-                                delta=ChatCompletionChunkDelta(
-                                    tool_calls=event.tool_calls,
-                                ),
-                                finish_reason=event.finish_reason,
-                            )
-                        ],
-                        usage=get_usage(output) if output.finished else None,
-                    )
-                    _tc_sse = f"data: {chunk.model_dump_json(exclude_none=True)}\n\n"
-                    logger.info(f"[SSE-TC] {_tc_sse.strip()[:300]}")
-                    yield _tc_sse
-
-                elif event.type == "finish":
-                    chunk = ChatCompletionChunk(
-                        id=response_id,
-                        model=_resolve_model_name(request.model),
-                        choices=[
-                            ChatCompletionChunkChoice(
-                                delta=ChatCompletionChunkDelta(
-                                    content=event.content,
-                                    reasoning_content=event.reasoning,
-                                ),
-                                finish_reason=event.finish_reason,
-                                logprobs=_build_chunk_logprobs(output),
-                            )
-                        ],
-                        usage=get_usage(output) if output.finished else None,
-                    )
-                    yield f"data: {chunk.model_dump_json(exclude_none=True)}\n\n"
-
-        # Fallback tool call detection (closing tag never arrived, etc.)
-        for event in processor.finalize():
-            if event.type == "tool_call":
-                tool_chunk = ChatCompletionChunk(
-                    id=response_id,
-                    model=_resolve_model_name(request.model),
-                    choices=[
-                        ChatCompletionChunkChoice(
-                            delta=ChatCompletionChunkDelta(
-                                tool_calls=event.tool_calls,
-                            ),
-                            finish_reason="tool_calls",
-                        )
-                    ],
-                )
-                _fb_sse = f"data: {tool_chunk.model_dump_json(exclude_none=True)}\n\n"
-                logger.info(f"[SSE-FALLBACK-TC] {_fb_sse.strip()[:300]}")
-                yield _fb_sse
-
-        # Log throughput
-        elapsed = time.perf_counter() - start_time
-        tokens_per_sec = completion_tokens / elapsed if elapsed > 0 else 0
-        logger.info(
-            f"Chat completion (stream): {completion_tokens} tokens in {elapsed:.2f}s ({tokens_per_sec:.1f} tok/s)"
-        )
-
-        # Send final chunk with usage if requested
-        if include_usage:
-            usage_chunk = ChatCompletionChunk(
-                id=response_id,
-                model=_resolve_model_name(request.model),
-                choices=[],
-                usage=Usage(
-                    prompt_tokens=prompt_tokens,
-                    completion_tokens=completion_tokens,
-                    total_tokens=prompt_tokens + completion_tokens,
-                ),
-            )
-            yield f"data: {usage_chunk.model_dump_json(exclude_none=True)}\n\n"
-
-        yield "data: [DONE]\n\n"
-    finally:
-        if cfg.gc_control and gc_was_enabled:
-            gc.enable()
-            gc.collect()
+# Streaming Helpers — moved to routes/completions.py and routes/chat.py
 
 
 # =============================================================================
@@ -2779,7 +793,10 @@ async def init_mcp(config_path: str):
 # Route modules — imported after all server globals are defined to avoid
 # circular imports (route modules import verify_api_key etc. from this module)
 # =============================================================================
+from .routes.anthropic import router as _anthropic_router
 from .routes.audio import router as _audio_router
+from .routes.chat import router as _chat_router
+from .routes.completions import router as _completions_router
 from .routes.embeddings import router as _embeddings_router
 from .routes.health import router as _health_router
 from .routes.mcp_routes import router as _mcp_router
@@ -2787,6 +804,9 @@ from .routes.models import router as _models_router
 
 app.include_router(_health_router)
 app.include_router(_models_router)
+app.include_router(_chat_router)
+app.include_router(_completions_router)
+app.include_router(_anthropic_router)
 app.include_router(_embeddings_router)
 app.include_router(_mcp_router)
 app.include_router(_audio_router)
