+++ vllm_mlx/service/__init__.py
@@ -1,1 +1,2 @@
 # SPDX-License-Identifier: Apache-2.0
+"""Service layer — shared helpers and post-processing pipeline."""
